Renomeando e reorganizando componentes de ações, extraindo lógica de exportação para AcoesEscala e ajustando estilos.
