+++ front/src/components/acoes/styles.module.css
@@ -1,4 +1,4 @@
-.acoesEscala {
+.acoes {
     display: flex;
     flex-direction: column;
     justify-content: start;
@@ -6,13 +6,13 @@
     margin-bottom: 20px;
 }
 
-.acoesEscala div {
+.acoes div {
     display: flex;
     align-items: center;
     gap: 10px;
 }
 
-.acoesEscala button {
+.acoes button {
     padding: 8px 20px;
     background-color: #ef5c41;
     border: none;
@@ -25,6 +25,6 @@
     align-items: center;
 }
 
-.acoesEscala button:hover {
+.acoes button:hover {
     background-color: rgba(255, 99, 71, 0.625);
 }
